requirements.txt
streamlit
requests
python-binance
